Create HTMLInputElement backing field if imeOptions.singeLine
